requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
charset-normalizer==3.4.6
click==8.3.1
fastapi==0.135.2
h11==0.16.0
idna==3.11
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pypinyin==0.55.0
PyYAML==6.0.3
reportlab==4.4.10
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
